genai/requirements.txt
docling==2.36.1
fastapi==0.115.12
langchain==0.3.25
langchain_core==0.3.65
langchain_docling==1.0.0
pydantic==2.11.5
python-dotenv==1.1.0
Requests==2.32.4
